requirements.txt.txt
opencv-python
ultralytics
torch
torchvision
torchaudio
requests
pillow
io
Flask
json
tkinter
customtkinter
face_recognition
numpy
shutil
mysql-connector-python